+++ R/magma_read_sets_out.R
@@ -7,25 +7,23 @@
 #' 
 #' @keywords internal
 magma_read_sets_out <- function(out_prefix,
-                                verbose = TRUE){ 
-    # version <- magma_installed_version(verbose = FALSE)
-    # v <- as.numeric(gsub(paste(letters,collapse = "|"),"",version))
-    # if(v < 1.09){
-    messager("Reading sets out info from .set.out file.",
-             v=verbose)
-    path <- sprintf("%s.sets.out",out_prefix)
-    res <- utils::read.table(file = path,
-                              header = TRUE,
-                              stringsAsFactors = FALSE) 
-    # } else {
-    #     messager("Reading sets out info from .gsa.out file.",
-    #              v=verbose)
-    #     #### Not totally sure if this is right 
-    #     path <- sprintf("%s.gsa.out", out_prefix)
-    #     res <- utils::read.table(file = path, 
-    #                               header = TRUE,
-    #                               stringsAsFactors = FALSE) 
-    #     res <- res[res$TYPE == "COVAR" & (res$MODEL == 1), ] 
-    # }
+                                verbose = TRUE){  
+    sets.out <- sprintf("%s.sets.out",out_prefix)
+    if(file.exists(sets.out)){
+        messager("Reading sets out info from .set.out file.",
+                 v=verbose) 
+        res <- utils::read.table(file = sets.out,
+                                  header = TRUE,
+                                  stringsAsFactors = FALSE) 
+    } else {
+        messager("Reading sets out info from .gsa.out file.",
+                 v=verbose)
+        #### Not totally sure if this is right
+        path <- sprintf("%s.gsa.out", out_prefix)
+        res <- utils::read.table(file = path,
+                                  header = TRUE,
+                                  stringsAsFactors = FALSE)
+        res <- res[(res$TYPE == "SET") & (res$MODEL == 1), ]
+    }
     return(res)
 }